backend/requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
